test: split the repo-init marker wait out of hosted-session.test.ts
The file crossed the 500-line cap. The seam is real: the rest of it is about
the paste itself (does the prompt reach the composer, do the busy gates hold),
while these four are about one file on disk and what its contents mean, and
they own temp-dir plumbing nothing else there needs.
